Refactor model classes for named exports.
