+++ src/model/index.js
@@ -1,0 +1,1 @@
+export * from './Quest.js';
